SONAR-6831 move queue related classes to their own package
