+++ server/sonar-server/src/main/java/org/sonar/server/computation/batch/BatchReportDirectoryHolder.java
@@ -20,10 +20,11 @@
 package org.sonar.server.computation.batch;
 
 import java.io.File;
+import org.sonar.server.computation.queue.CeTask;
 
 public interface BatchReportDirectoryHolder {
   /**
-   * The File of the directory where the Batch report files for the current {@link org.sonar.server.computation.CeTask} are stored.
+   * The File of the directory where the Batch report files for the current {@link CeTask} are stored.
    *
    * @throws IllegalStateException if the holder is empty (ie. there is no directory yet)
    */
