+++ server/sonar-server/src/main/java/org/sonar/server/computation/queue/CeProcessingSchedulerExecutorService.java
@@ -17,20 +17,12 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.server.computation;
+package org.sonar.server.computation.queue;
 
-import org.sonar.db.ce.CeActivityDto;
+import org.sonar.server.util.StoppableScheduledExecutorService;
 
-public class CleanReportQueueListener implements CeQueueListener {
-
-  private final ReportFiles reportFiles;
-
-  public CleanReportQueueListener(ReportFiles reportFiles) {
-    this.reportFiles = reportFiles;
-  }
-
-  @Override
-  public void onRemoved(CeTask task, CeActivityDto.Status status) {
-    reportFiles.deleteIfExists(task.getUuid());
-  }
+/**
+ * The {@link java.util.concurrent.ExecutorService} responsible for running {@link CeWorkerRunnable}.
+ */
+public interface CeProcessingSchedulerExecutorService extends StoppableScheduledExecutorService {
 }
