+++ server/sonar-server/src/main/java/org/sonar/server/computation/queue/CeProcessingSchedulerImpl.java
@@ -17,9 +17,10 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.server.computation;
+package org.sonar.server.computation.queue;
 
 import java.util.concurrent.TimeUnit;
+import org.sonar.server.computation.queue.report.ReportTaskProcessor;
 import org.sonar.server.computation.log.CeLogging;
 
 public class CeProcessingSchedulerImpl implements CeProcessingScheduler {
