+++ server/sonar-server/src/main/java/org/sonar/server/computation/queue/CeQueueImpl.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.server.computation;
+package org.sonar.server.computation.queue;
 
 import com.google.common.base.Optional;
 import java.util.concurrent.atomic.AtomicBoolean;
