+++ server/sonar-server/src/main/java/org/sonar/server/computation/queue/CeWorkerRunnable.java
@@ -18,7 +18,7 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-package org.sonar.server.computation;
+package org.sonar.server.computation.queue;
 
 import com.google.common.base.Optional;
 import org.sonar.api.utils.log.Logger;
@@ -26,6 +26,7 @@ import org.sonar.api.utils.log.Loggers;
 import org.sonar.core.util.logs.Profiler;
 import org.sonar.db.ce.CeActivityDto;
 import org.sonar.server.computation.log.CeLogging;
+import org.sonar.server.computation.queue.report.ReportTaskProcessor;
 
 import static java.lang.String.format;
