+++ server/sonar-server/src/main/java/org/sonar/server/computation/queue/report/CleanReportQueueListener.java
@@ -17,10 +17,23 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.server.computation;
+package org.sonar.server.computation.queue.report;
 
-public interface CeProcessingScheduler {
+import org.sonar.db.ce.CeActivityDto;
+import org.sonar.server.computation.ReportFiles;
+import org.sonar.server.computation.queue.CeQueueListener;
+import org.sonar.server.computation.queue.CeTask;
 
-  void startScheduling();
+public class CleanReportQueueListener implements CeQueueListener {
 
+  private final ReportFiles reportFiles;
+
+  public CleanReportQueueListener(ReportFiles reportFiles) {
+    this.reportFiles = reportFiles;
+  }
+
+  @Override
+  public void onRemoved(CeTask task, CeActivityDto.Status status) {
+    reportFiles.deleteIfExists(task.getUuid());
+  }
 }
