+++ server/sonar-server/src/main/java/org/sonar/server/computation/queue/report/ReportTaskProcessor.java
@@ -17,11 +17,13 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.server.computation;
+package org.sonar.server.computation.queue.report;
 
 import org.sonar.core.platform.ComponentContainer;
+import org.sonar.server.computation.ComputationStepExecutor;
 import org.sonar.server.computation.container.ComputeEngineContainer;
 import org.sonar.server.computation.container.ContainerFactory;
+import org.sonar.server.computation.queue.CeTask;
 
 public class ReportTaskProcessor {
