+++ server/sonar-server/src/main/java/org/sonar/server/computation/ws/CancelAllWsAction.java
@@ -23,7 +23,7 @@ import org.sonar.api.server.ws.Request;
 import org.sonar.api.server.ws.Response;
 import org.sonar.api.server.ws.WebService;
 import org.sonar.api.web.UserRole;
-import org.sonar.server.computation.CeQueue;
+import org.sonar.server.computation.queue.CeQueue;
 import org.sonar.server.user.UserSession;
 
 public class CancelAllWsAction implements CeWsAction {
