+++ server/sonar-server/src/main/java/org/sonar/server/computation/ws/SubmitWsAction.java
@@ -24,8 +24,8 @@ import org.apache.commons.lang.StringUtils;
 import org.sonar.api.server.ws.Request;
 import org.sonar.api.server.ws.Response;
 import org.sonar.api.server.ws.WebService;
-import org.sonar.server.computation.CeTask;
-import org.sonar.server.computation.ReportSubmitter;
+import org.sonar.server.computation.queue.CeTask;
+import org.sonar.server.computation.queue.report.ReportSubmitter;
 import org.sonar.server.ws.WsUtils;
 import org.sonarqube.ws.WsCe;
