+++ server/sonar-server/src/main/java/org/sonar/server/platform/platformlevel/PlatformLevel4.java
@@ -62,13 +62,13 @@ import org.sonar.server.component.DefaultRubyComponentService;
 import org.sonar.server.component.ws.ComponentsWs;
 import org.sonar.server.component.ws.EventsWs;
 import org.sonar.server.component.ws.ResourcesWs;
-import org.sonar.server.computation.CeQueueCleaner;
-import org.sonar.server.computation.CeQueueImpl;
-import org.sonar.server.computation.CeQueueInitializer;
-import org.sonar.server.computation.CleanReportQueueListener;
+import org.sonar.server.computation.queue.CeQueueImpl;
+import org.sonar.server.computation.queue.CeQueueCleaner;
+import org.sonar.server.computation.queue.CeQueueInitializer;
+import org.sonar.server.computation.queue.report.CleanReportQueueListener;
 import org.sonar.server.computation.ComputeEngineProcessingModule;
 import org.sonar.server.computation.ReportFiles;
-import org.sonar.server.computation.ReportSubmitter;
+import org.sonar.server.computation.queue.report.ReportSubmitter;
 import org.sonar.server.computation.dbcleaner.IndexPurgeListener;
 import org.sonar.server.computation.dbcleaner.ProjectCleaner;
 import org.sonar.server.computation.log.CeLogging;
