+++ server/sonar-server/src/test/java/org/sonar/server/computation/queue/CeQueueCleanerTest.java
@@ -17,29 +17,33 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.server.computation;
+package org.sonar.server.computation.queue;
 
 import java.io.File;
 import java.io.IOException;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
-import org.mockito.InOrder;
 import org.mockito.Mockito;
+import org.sonar.api.platform.ServerUpgradeStatus;
 import org.sonar.api.utils.System2;
 import org.sonar.db.DbSession;
 import org.sonar.db.DbTester;
 import org.sonar.db.ce.CeQueueDto;
 import org.sonar.db.ce.CeTaskTypes;
-import org.sonar.server.computation.monitoring.CEQueueStatus;
-import org.sonar.server.computation.monitoring.CEQueueStatusImpl;
+import org.sonar.server.computation.ReportFiles;
+import org.sonar.server.computation.queue.CeQueueCleaner;
+import org.sonar.server.computation.queue.CeQueueImpl;
 
+import static java.util.Arrays.asList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Matchers.any;
+import static org.mockito.Matchers.eq;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
-public class CeQueueInitializerTest {
+public class CeQueueCleanerTest {
 
   @Rule
   public DbTester dbTester = DbTester.create(System2.INSTANCE);
@@ -47,39 +51,49 @@ public class CeQueueInitializerTest {
   @Rule
   public TemporaryFolder tempFolder = new TemporaryFolder();
 
+  ServerUpgradeStatus serverUpgradeStatus = mock(ServerUpgradeStatus.class);
   ReportFiles reportFiles = mock(ReportFiles.class, Mockito.RETURNS_DEEP_STUBS);
-  CEQueueStatus queueStatus = new CEQueueStatusImpl();
-  CeQueueCleaner cleaner = mock(CeQueueCleaner.class);
-  CeProcessingScheduler scheduler = mock(CeProcessingScheduler.class);
-  CeQueueInitializer underTest = new CeQueueInitializer(dbTester.getDbClient(), queueStatus, cleaner, scheduler);
+  CeQueueImpl queue = mock(CeQueueImpl.class);
+  CeQueueCleaner underTest = new CeQueueCleaner(dbTester.getDbClient(), serverUpgradeStatus, reportFiles, queue);
 
   @Test
-  public void init_jmx_counters() throws IOException {
+  public void reset_in_progress_tasks_to_pending() throws IOException {
     insertInQueue("TASK_1", CeQueueDto.Status.PENDING);
-    insertInQueue("TASK_2", CeQueueDto.Status.PENDING);
-    // this in-progress task is going to be moved to PENDING
-    insertInQueue("TASK_3", CeQueueDto.Status.IN_PROGRESS);
+    insertInQueue("TASK_2", CeQueueDto.Status.IN_PROGRESS);
 
-    underTest.start();
+    underTest.clean(dbTester.getSession());
 
-    assertThat(queueStatus.getPendingCount()).isEqualTo(3);
+    assertThat(dbTester.getDbClient().ceQueueDao().countByStatus(dbTester.getSession(), CeQueueDto.Status.PENDING)).isEqualTo(2);
+    assertThat(dbTester.getDbClient().ceQueueDao().countByStatus(dbTester.getSession(), CeQueueDto.Status.IN_PROGRESS)).isEqualTo(0);
   }
 
   @Test
-  public void init_jmx_counters_when_queue_is_empty() {
-    underTest.start();
+  public void clear_queue_if_version_upgrade() {
+    when(serverUpgradeStatus.isUpgraded()).thenReturn(true);
 
-    assertThat(queueStatus.getPendingCount()).isEqualTo(0);
+    underTest.clean(dbTester.getSession());
+
+    verify(queue).clear();
   }
 
   @Test
-  public void clean_queue_then_start_scheduler_of_workers() throws IOException {
-    InOrder inOrder = Mockito.inOrder(cleaner, scheduler);
+  public void cancel_task_if_report_file_is_missing() throws IOException {
+    CeQueueDto task = insertInQueue("TASK_1", CeQueueDto.Status.PENDING, false);
 
-    underTest.start();
+    underTest.clean(dbTester.getSession());
 
-    inOrder.verify(cleaner).clean(any(DbSession.class));
-    inOrder.verify(scheduler).startScheduling();
+    verify(queue).cancel(any(DbSession.class), eq(task));
+  }
+
+  @Test
+  public void delete_orphan_report_files() throws Exception {
+    // two files on disk but on task in queue
+    insertInQueue("TASK_1", CeQueueDto.Status.PENDING, true);
+    when(reportFiles.listUuids()).thenReturn(asList("TASK_1", "TASK_2"));
+
+    underTest.clean(dbTester.getSession());
+
+    verify(reportFiles).deleteIfExists("TASK_2");
   }
 
   private void insertInQueue(String taskUuid, CeQueueDto.Status status) throws IOException {
