+++ server/sonar-server/src/test/java/org/sonar/server/computation/queue/CeWorkerRunnableTest.java
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software Foundation,
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
-package org.sonar.server.computation;
+package org.sonar.server.computation.queue;
 
 import com.google.common.base.Optional;
 import org.junit.Test;
@@ -25,6 +25,7 @@ import org.mockito.InOrder;
 import org.mockito.Mockito;
 import org.sonar.db.ce.CeActivityDto;
 import org.sonar.db.ce.CeTaskTypes;
+import org.sonar.server.computation.queue.report.ReportTaskProcessor;
 import org.sonar.server.computation.log.CeLogging;
 
 import static org.mockito.Mockito.doThrow;
