+++ server/sonar-server/src/test/java/org/sonar/server/computation/ws/SubmitWsActionTest.java
@@ -24,8 +24,8 @@ import org.junit.Test;
 import org.mockito.Matchers;
 import org.sonar.core.util.Protobuf;
 import org.sonar.db.ce.CeTaskTypes;
-import org.sonar.server.computation.CeTask;
-import org.sonar.server.computation.ReportSubmitter;
+import org.sonar.server.computation.queue.CeTask;
+import org.sonar.server.computation.queue.report.ReportSubmitter;
 import org.sonar.server.plugins.MimeTypes;
 import org.sonar.server.ws.TestResponse;
 import org.sonar.server.ws.WsActionTester;
